% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_int.R
\name{find_int}
\alias{find_int}
\title{Title}
\usage{
find_int(mean_in, mean, sd, n)
}
\arguments{
\item{mean_in}{- the mean to be confidence tested}

\item{mean}{- mean of the normal sample}

\item{sd}{- sd of normal sample}

\item{n}{-  size of sample}
}
\description{
Title
}
