% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Fire Distance}
  \item{DAMAGE}{Damage done by fire}
}
}
\source{
FIREDAM.csv
}
\usage{
fire
}
\description{
A report that details the damage from fires.
}
\keyword{datasets}
