% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearFitPlot.R
\name{LinearFitPlot}
\alias{LinearFitPlot}
\title{LinearFitPlot}
\usage{
LinearFitPlot(xData, yData, title, xaxislab, yaxislab)
}
\arguments{
\item{xData}{- The x axis data}

\item{yData}{- The y axis data}

\item{title}{- The title name (String)}

\item{xaxislab}{- x axis label (String)}

\item{yaxislab}{- y axis label (String)}
}
\value{
data.lm - the linear model of the dataset
}
\description{
This function creates
}
