% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{- The number of times the experiment is run}

\item{n}{- The number of samples}

\item{p}{- The probability of a success in each sample}
}
\description{
mybin function
}
