% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n = 5, iter = 1000, lambda = 10, ...)
}
\arguments{
\item{n}{- the number of samples in a trial}

\item{iter}{- the number of trials}

\item{lambda}{- lambda}

\item{...}{- others}
}
\description{
mycltp
}
\examples{
mycltp(n= 5,iter = 1000,lambda=10)
}
