% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{- the function to be plotted}

\item{x}{- list of samples}

\item{param}{- x axis}

\item{...}{- other params}
}
\value{
- list with parameters
}
\description{
mymaxlik
}
