% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{- the mean of the normal distribution}

\item{sigma}{- the standard deviation of the normal distribution}

\item{a}{- the value below which to calculate the lower-tail probability}
}
\value{
- a list containing relavent variables and the probabilty of left tail distribution of a
}
\description{
Title
}
