% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{- the number of seats on the flight}

\item{gamma}{- the accepted probability that a flight will be overbooked}

\item{p}{- probability of a passenger showing}
}
\description{
This function takes a number of seats N, a tolerated probability of a overbooking gamma, and
probability of passengers showing p, to find the discrete and continuous optimized number of
tickets to sell.
}
\author{
Jonathan Douglas
}
